Rename : 파일 위치 이동 및 호출 경로 수정
[수정된 파일]
validation.test.js : 폴더 이동에 따른 호출 경로를 수정을 하였습니다.
- validation.js

Race.js : 폴더 이동에 따른 호출 경로를 수정을 하였습니다.
- random.js

Input.js :  폴더 이동에 따른 호출 경로를 수정을 하였습니다.
- readLine.js

Output.js :  폴더 이동에 따른 호출 경로를 수정을 하였습니다.
- positionToStr.js

RaceSetup.js :  폴더 이동에 따른 호출 경로를 수정을 하였습니다.
- tryInput.js
- validation.js

tryInput.js :  폴더 이동 및 폴더 이동에 따른 호출 경로를 수정을 하였습니다. (view/ -> utils/)
- errorHandler.js

errorHandler.js : 파일이 이동되었습니다. (view/ -> utils/)
positionToStr.js : 파일이 이동되었습니다. (domain/ -> utils/)
random.js : 파일이 이동되었습니다. (domain/ -> utils/)
readLine.js : 파일이 이동되었습니다. (view/ -> utils/)
validation.js : 파일이 이동되었습니다. (domain/ -> validation/)
