+++ src/view/Output.js
@@ -1,4 +1,4 @@
-import positionToStr from '../domain/positionToStr.js';
+import positionToStr from '../utils/positionToStr.js';
 import OUTPUT from '../constant/output.js';
 
 const OutputView = {
